Improved modularity with a custom hook for managing list navigation with keyboard.
